Added Functionality to KitchenPanel so that it will populate the table.
OrderChunk Class - added a get and set method as well as an OrderID
number 
that should come from the Database.

Created a new Enum for status - "FINISHED"

ResturantSystem Class - added a method that needs to be completed which 
accesses the database and returns an arraylist which contains all the 
OrderChunks which do not have their status set to "out" or "finished".  
Hopefully the returned list will be in ascending order and will also
assign 
OrderID's to the OrderChunks